#24 Reafactor: 도메인 위치 수정
