--- src/main/java/com/app/interstory/episode/repository/PointRepository.java
@@ -1,12 +1,0 @@
-package com.app.interstory.episode.repository;
-
-import java.util.List;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.app.interstory.user.domain.entity.Point;
-
-public interface PointRepository extends JpaRepository<Point, Long> {
-	List<Point> findByUser(Long userId);
-}
-
