+++ src/main/java/com/app/interstory/novel/dto/response/EpisodeResponseDTO.java
@@ -1,17 +1,24 @@
-package com.app.interstory.episode.domain.dto;
+package com.app.interstory.novel.dto.response;
 
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+
+import java.sql.Timestamp;
 
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class EpisodeRequestDTO {
+@Builder
+public class EpisodeResponseDTO {
+	private Long episodeId;
 	private Long novelId;
 	private String title;
-	private String thumbnailRenamedFilename;
+	private Integer viewCount;
+	private Timestamp publishedAt;
 	private String thumbnailUrl;
+	private Integer likeCount;
 	private String content;
 	private Boolean status;
 }
