+++ src/main/java/com/app/interstory/novel/repository/CartItemRepository.java
@@ -1,11 +1,11 @@
-package com.app.interstory.episode.repository;
+package com.app.interstory.novel.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.app.interstory.novel.domain.entity.Collection;
 import com.app.interstory.novel.domain.entity.Episode;
+import com.app.interstory.user.domain.entity.CartItem;
 import com.app.interstory.user.domain.entity.User;
 
-public interface CollectionRepository extends JpaRepository<Collection, Long> {
+public interface CartItemRepository extends JpaRepository<CartItem, Long> {
 	boolean existsByUserAndEpisode(User user, Episode episode);
 }
